Factored out grppi::optional
